Refactored copy link button to its own component file, change link text upon copy. Removed from dataFormRow
